finnhub-pipeline/finnhub_producer/requirements.txt
kafka-python==2.0.2
websocket-client==1.6.4
prometheus-client==0.19.0
python-dotenv==1.0.0